+++ app/app.go
@@ -44,6 +44,7 @@ import (
 	"github.com/binance-chain/node/plugins/param/paramhub"
 	"github.com/binance-chain/node/plugins/tokens"
 	tkstore "github.com/binance-chain/node/plugins/tokens/store"
+	"github.com/binance-chain/node/plugins/tokens/swap"
 	"github.com/binance-chain/node/plugins/tokens/timelock"
 	"github.com/binance-chain/node/wire"
 )
@@ -84,13 +85,15 @@ type BinanceChain struct {
 	stakeKeeper    stake.Keeper
 	govKeeper      gov.Keeper
 	timeLockKeeper timelock.Keeper
+	swapKeeper     swap.Keeper
 	// keeper to process param store and update
 	ParamHub *param.ParamHub
 
-	baseConfig        *config.BaseConfig
-	upgradeConfig     *config.UpgradeConfig
-	publicationConfig *config.PublicationConfig
-	publisher         pub.MarketDataPublisher
+	baseConfig         *config.BaseConfig
+	upgradeConfig      *config.UpgradeConfig
+	abciQueryBlackList map[string]bool
+	publicationConfig  *config.PublicationConfig
+	publisher          pub.MarketDataPublisher
 
 	// Unlike tendermint, we don't need implement a no-op metrics, usage of this field should
 	// check nil-ness to know whether metrics collection is turn on
@@ -109,12 +112,13 @@ func NewBinanceChain(logger log.Logger, db dbm.DB, traceStore io.Writer, baseApp
 
 	// create the applicationsimulate object
 	var app = &BinanceChain{
-		BaseApp:           baseapp.NewBaseApp(appName /*, cdc*/, logger, db, decoders, sdk.CollectConfig{ServerContext.PublishAccountBalance, ServerContext.PublishTransfer}, baseAppOptions...),
-		Codec:             cdc,
-		queryHandlers:     make(map[string]types.AbciQueryHandler),
-		baseConfig:        ServerContext.BaseConfig,
-		upgradeConfig:     ServerContext.UpgradeConfig,
-		publicationConfig: ServerContext.PublicationConfig,
+		BaseApp:            baseapp.NewBaseApp(appName /*, cdc*/, logger, db, decoders, sdk.CollectConfig{ServerContext.PublishAccountBalance, ServerContext.PublishTransfer}, baseAppOptions...),
+		Codec:              cdc,
+		queryHandlers:      make(map[string]types.AbciQueryHandler),
+		baseConfig:         ServerContext.BaseConfig,
+		upgradeConfig:      ServerContext.UpgradeConfig,
+		abciQueryBlackList: getABCIQueryBlackList(ServerContext.QueryConfig),
+		publicationConfig:  ServerContext.PublicationConfig,
 	}
 	// set upgrade config
 	SetUpgradeConfig(app.upgradeConfig)
@@ -148,6 +152,8 @@ func NewBinanceChain(logger log.Logger, db dbm.DB, traceStore io.Writer, baseApp
 	app.timeLockKeeper = timelock.NewKeeper(cdc, common.TimeLockStoreKey, app.CoinKeeper, app.AccountKeeper,
 		timelock.DefaultCodespace)
 
+	app.swapKeeper = swap.NewKeeper(cdc, common.AtomicSwapStoreKey, app.CoinKeeper, app.Pool, swap.DefaultCodespace)
+
 	// legacy bank route (others moved to plugin init funcs)
 	app.Router().
 		AddRoute("bank", bank.NewHandler(app.CoinKeeper)).
@@ -157,6 +163,7 @@ func NewBinanceChain(logger log.Logger, db dbm.DB, traceStore io.Writer, baseApp
 	app.QueryRouter().AddRoute("gov", gov.NewQuerier(app.govKeeper))
 	app.QueryRouter().AddRoute("stake", stake.NewQuerier(app.stakeKeeper, cdc))
 	app.QueryRouter().AddRoute("timelock", timelock.NewQuerier(app.timeLockKeeper))
+	app.QueryRouter().AddRoute(swap.AtomicSwapRoute, swap.NewQuerier(app.swapKeeper))
 
 	app.RegisterQueryHandler("account", app.AccountHandler)
 	app.RegisterQueryHandler("admin", admin.GetHandler(ServerContext.Config))
@@ -207,6 +214,7 @@ func NewBinanceChain(logger log.Logger, db dbm.DB, traceStore io.Writer, baseApp
 		common.StakeStoreKey,
 		common.GovStoreKey,
 		common.TimeLockStoreKey,
+		common.AtomicSwapStoreKey,
 	)
 	app.SetAnteHandler(tx.NewAnteHandler(app.AccountKeeper))
 	app.SetPreChecker(tx.NewTxPreChecker())
@@ -250,9 +258,11 @@ func SetUpgradeConfig(upgradeConfig *config.UpgradeConfig) {
 	upgrade.Mgr.AddUpgradeHeight(upgrade.BEP10, upgradeConfig.BEP10Height)
 	upgrade.Mgr.AddUpgradeHeight(upgrade.BEP12, upgradeConfig.BEP12Height)
 	upgrade.Mgr.AddUpgradeHeight(upgrade.BEP19, upgradeConfig.BEP19Height)
+	upgrade.Mgr.AddUpgradeHeight(upgrade.BEP3, upgradeConfig.BEP3Height)
 
 	// register store keys of upgrade
 	upgrade.Mgr.RegisterStoreKeys(upgrade.BEP9, common.TimeLockStoreKey.Name())
+	upgrade.Mgr.RegisterStoreKeys(upgrade.BEP3, common.AtomicSwapStoreKey.Name())
 
 	// register msg types of upgrade
 	upgrade.Mgr.RegisterMsgTypes(upgrade.BEP9,
@@ -263,6 +273,20 @@ func SetUpgradeConfig(upgradeConfig *config.UpgradeConfig) {
 
 	// register msg types of upgrade
 	upgrade.Mgr.RegisterMsgTypes(upgrade.BEP12, account.SetAccountFlagsMsg{}.Type())
+	upgrade.Mgr.RegisterMsgTypes(upgrade.BEP3,
+		swap.HTLTMsg{}.Type(),
+		swap.DepositHTLTMsg{}.Type(),
+		swap.ClaimHTLTMsg{}.Type(),
+		swap.RefundHTLTMsg{}.Type(),
+	)
+}
+
+func getABCIQueryBlackList(queryConfig *config.QueryConfig) map[string]bool {
+	cfg := make(map[string]bool)
+	for _, path := range queryConfig.ABCIQueryBlackList {
+		cfg[path] = true
+	}
+	return cfg
 }
 
 func (app *BinanceChain) initRunningMode() {
@@ -300,7 +324,7 @@ func (app *BinanceChain) initDex(pairMapper dex.TradingPairMapper) {
 }
 
 func (app *BinanceChain) initPlugins() {
-	tokens.InitPlugin(app, app.TokenMapper, app.AccountKeeper, app.CoinKeeper, app.timeLockKeeper)
+	tokens.InitPlugin(app, app.TokenMapper, app.AccountKeeper, app.CoinKeeper, app.timeLockKeeper, app.swapKeeper)
 	dex.InitPlugin(app, app.DexKeeper, app.TokenMapper, app.AccountKeeper, app.govKeeper)
 	param.InitPlugin(app, app.ParamHub)
 	account.InitPlugin(app, app.AccountKeeper)
@@ -505,6 +529,7 @@ func (app *BinanceChain) EndBlocker(ctx sdk.Context, req abci.RequestEndBlock) a
 		icoDone := ico.EndBlockAsync(ctx)
 		dex.EndBreatheBlock(ctx, app.DexKeeper, app.govKeeper, height, blockTime)
 		param.EndBreatheBlock(ctx, app.ParamHub)
+		tokens.EndBreatheBlock(ctx, app.swapKeeper)
 		// other end blockers
 		<-icoDone
 	} else {
@@ -610,6 +635,10 @@ func (app *BinanceChain) Query(req abci.RequestQuery) (res abci.ResponseQuery) {
 	path := baseapp.SplitPath(req.Path)
 	if len(path) == 0 {
 		msg := "no query path provided"
+		return sdk.ErrUnknownRequest(msg).QueryResult()
+	}
+	if app.abciQueryBlackList[req.Path] {
+		msg := fmt.Sprintf("abci query interface (%s) is in black list", req.Path)
 		return sdk.ErrUnknownRequest(msg).QueryResult()
 	}
 	prefix := path[0]