+++ common/upgrade/upgrade.go
@@ -7,13 +7,17 @@ var Mgr = sdk.UpgradeMgr
 // prefix for the upgrade name
 // bugfix: fix
 // improvement: (maybe bep ?)
-
-// Galileo Upgrade
-const BEP6 = "BEP6"   // https://github.com/binance-chain/BEPs/pull/6
-const BEP9 = "BEP9"   // https://github.com/binance-chain/BEPs/pull/9
-const BEP10 = "BEP10" // https://github.com/binance-chain/BEPs/pull/10
-const BEP19 = "BEP19" // https://github.com/binance-chain/BEPs/pull/19  match engine revision
-const BEP12 = "BEP12" // https://github.com/binance-chain/BEPs/pull/17
+const (
+	// Galileo Upgrade
+	BEP6  = "BEP6"  // https://github.com/binance-chain/BEPs/pull/6
+	BEP9  = "BEP9"  // https://github.com/binance-chain/BEPs/pull/9
+	BEP10 = "BEP10" // https://github.com/binance-chain/BEPs/pull/10
+	BEP19 = "BEP19" // https://github.com/binance-chain/BEPs/pull/19  match engine revision
+	// Hubble Upgrade
+	BEP12 = "BEP12" // https://github.com/binance-chain/BEPs/pull/17
+	// Archimedes Upgrade
+	BEP3 = "BEP3" // https://github.com/binance-chain/BEPs/pull/30
+)
 
 func UpgradeBEP10(before func(), after func()) {
 	sdk.Upgrade(BEP10, before, nil, after)