+++ integration_test.sh
@@ -28,6 +28,7 @@ function prepare_node() {
 	secret=$(./bnbchaind init --moniker testnode --home ${home} --home-client ${cli_home} --chain-id ${chain_id} | grep secret | grep -o ":.*" | grep -o "\".*"  | sed "s/\"//g")
 
     $(cd "./${home}/config" && sed -i -e "s/BEP12Height = 9223372036854775807/BEP12Height = 1/g" app.toml)
+    $(cd "./${home}/config" && sed -i -e "s/BEP3Height = 9223372036854775807/BEP3Height = 1/g" app.toml)
 	$(cd "./${home}/config" && sed -i -e "s/skip_timeout_commit = false/skip_timeout_commit = true/g" config.toml)
 	$(cd "./${home}/config" && sed -i -e "s/log_level = \"main\:info,state\:info,\*\:error\"/log_level = \"*\:debug\"/g" config.toml)
 	$(cd "./${home}/config" && sed -i -e 's/"voting_period": "1209600000000000"/"voting_period": "5000000000"/g' genesis.json)
@@ -208,5 +209,110 @@ check_operation "Send Token" "${result}" "ERROR:"
 
 result=$(expect ./send.exp ${cli_home} alice ${chain_id} "100000000000000:BNB" ${bob_addr} "1234567890")
 check_operation "Send Token" "${result}" "${chain_operation_words}"
+
+
+## ROUND 4 ##
+sleep 1s
+# Create an atomic swap
+result=$(expect ./HTLT-cross-chain.exp 2000 "100000000:BNB" "100000000:BNB" $bob_addr 0xf2fbB6C41271064613D6f44C7EE9A6c471Ec9B25 alice ${chain_id} ${cli_home})
+check_operation "Create an atomic swap" "${result}" "${chain_operation_words}"
+randomNumber=$(sed 's/Random number: //g' <<< $(echo "${result}" | grep -o "Random number: [0-9a-z]*"))
+timestamp=$(sed 's/Timestamp: //g' <<< $(echo "${result}" | grep -o "Timestamp: [0-9]*"))
+randomNumberHash=$(sed 's/Random number hash: //g' <<< $(echo "${result}" | grep -o "Random number hash: [0-9a-z]*"))
+swapID=$(sed 's/swapID: //g' <<< $(echo "${result}" | tail -n 1 | grep -o "swapID: [0-9a-z]*"))
+sleep 1s
+
+atimicSwap=$(./bnbcli token query-swap --swap-id ${swapID} --trust-node)
+swapFrom=$(echo "${atimicSwap}" | jq -r '.from')
+check_operation "Check swap creator address" $swapFrom $alice_addr
+swapTo=$(echo "${atimicSwap}" | jq -r '.to')
+check_operation "swap recipient address" $swapTo $bob_addr
+
+result=$(./bnbcli account bnb1wxeplyw7x8aahy93w96yhwm7xcq3ke4f8ge93u --trust-node)
+swapDeadAddrBalance=$(echo "${result}" | jq -r '.value.base.coins[0].amount')
+check_operation "the balance of swap dead address" $swapDeadAddrBalance "100000000"
+
+result=$(./bnbcli account $bob_addr --trust-node)
+balanceBobBeforeClaim=$(echo "${result}" | jq -r '.value.base.coins[0].amount')
+
+# Claim an atomic swap
+result=$(expect ./claim.exp ${swapID} $randomNumber alice ${chain_id} ${cli_home})
+check_operation "claim an atomic swap" "${result}" "${chain_operation_words}"
+
+sleep 1s
+
+result=$(./bnbcli account $bob_addr --trust-node)
+balanceBobAfterClaim=$(echo "${result}" | jq -r '.value.base.coins[0].amount')
+check_operation "Bob balance after claim swap" "$(expr $balanceBobAfterClaim - $balanceBobBeforeClaim)" "100000000"
+
+# Create an atomic swap
+result=$(expect ./HTLT-cross-chain.exp 2000 "100000000:BNB" "100000000:BNB" $alice_addr 0xf2fbB6C41271064613D6f44C7EE9A6c471Ec9B25 bob ${chain_id} ${cli_home})
+check_operation "Create an atomic swap" "${result}" "${chain_operation_words}"
+swapID=$(sed 's/swapID: //g' <<< $(echo "${result}" | tail -n 1 | grep -o "swapID: [0-9a-z]*"))
+
+sleep 1s
+
+# Refund an atomic swap
+result=$(expect ./refund.exp ${swapID} alice ${chain_id} ${cli_home})
+check_operation "refund an atomic swap which is still not expired" "${result}" "ERROR"
+
+sleep 1s
+
+result=$(./bnbcli account bnb1wxeplyw7x8aahy93w96yhwm7xcq3ke4f8ge93u --trust-node)
+swapDeadAddrBalance=$(echo "${result}" | jq -r '.value.base.coins[0].amount')
+check_operation "the balance of swap dead address" $swapDeadAddrBalance "100000000"
+
+# issue token
+result=$(expect ./issue.exp ETH Ethereum 1000000000000000 true bob ${chain_id} ${cli_home})
+eth_symbol=$(echo "${result}" | tail -n 1 | grep -o "ETH-[0-9A-Z]*")
+check_operation "Issue Token" "${result}" "${chain_operation_words}"
+
+sleep 1s
+# Create a single chain atomic swa
+result=$(expect ./HTLT-single-chain.exp 2000 "100000000:BNB" "10000:${eth_symbol}" $bob_addr alice ${chain_id} ${cli_home})
+check_operation "Create a single chain atomic swap" "${result}" "${chain_operation_words}"
+randomNumber=$(sed 's/Random number: //g' <<< $(echo "${result}" | grep -o "Random number: [0-9a-z]*"))
+timestamp=$(sed 's/Timestamp: //g' <<< $(echo "${result}" | grep -o "Timestamp: [0-9]*"))
+randomNumberHash=$(sed 's/Random number hash: //g' <<< $(echo "${result}" | grep -o "Random number hash: [0-9a-z]*"))
+swapID=$(sed 's/swapID: //g' <<< $(echo "${result}" | tail -n 1 | grep -o "swapID: [0-9a-z]*"))
+sleep 1s
+
+# Deposit to a single chain atomic swap
+result=$(expect ./deposit.exp ${swapID} "10000:${eth_symbol}" bob ${chain_id} ${cli_home})
+check_operation "Deposit to a single chain atomic swap" "${result}" "${chain_operation_words}"
+
+sleep 1s
+# claim a single chain atomic swap
+result=$(expect ./claim.exp ${swapID} ${randomNumber} alice ${chain_id} ${cli_home})
+check_operation "claim a single chain atomic swap" "${result}" "${chain_operation_words}"
+
+sleep 1s
+# Deposit to a single chain atomic swap
+result=$(expect ./deposit.exp ${swapID} "10000:${eth_symbol}" bob ${chain_id} ${cli_home})
+check_operation "Deposit to a closed single chain atomic swap" "${result}" "ERROR"
+
+sleep 1s
+# Create a single chain atomic swa
+result=$(expect ./HTLT-single-chain.exp 500 "100000000:BNB" "10000:${eth_symbol}" $bob_addr alice ${chain_id} ${cli_home})
+check_operation "Create a single chain atomic swap" "${result}" "${chain_operation_words}"
+randomNumber=$(sed 's/Random number: //g' <<< $(echo "${result}" | grep -o "Random number: [0-9a-z]*"))
+timestamp=$(sed 's/Timestamp: //g' <<< $(echo "${result}" | grep -o "Timestamp: [0-9]*"))
+randomNumberHash=$(sed 's/Random number hash: //g' <<< $(echo "${result}" | grep -o "Random number hash: [0-9a-z]*"))
+swapID=$(sed 's/swapID: //g' <<< $(echo "${result}" | tail -n 1 | grep -o "swapID: [0-9a-z]*"))
+
+sleep 1s
+# Deposit to a single chain atomic swap
+result=$(expect ./deposit.exp ${swapID} "10000:${eth_symbol}" bob ${chain_id}  ${cli_home})
+check_operation "Deposit to a single chain atomic swap" "${result}" "${chain_operation_words}"
+
+sleep 5s
+# refund a single chain atomic swap
+result=$(expect ./refund.exp ${swapID} alice ${chain_id} ${cli_home})
+check_operation "refund a single chain atomic swap" "${result}" "${chain_operation_words}"
+
+sleep 1s
+# Deposit to a single chain atomic swap
+result=$(expect ./deposit.exp ${swapID} "10000:${eth_symbol}" bob ${chain_id}  ${cli_home})
+check_operation "Deposit to a expired single chain atomic swap" "${result}" "ERROR"
 
 exit_test 0