+++ plugins/account/client/cli/account_flags.go
@@ -9,8 +9,6 @@ import (
 	"github.com/spf13/viper"
 
 	clientFlags "github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/context"
-	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
 
 	"github.com/binance-chain/node/common/client"
 	"github.com/binance-chain/node/common/types"
@@ -43,6 +41,21 @@ func setAccountFlagsCmd(cdc *wire.Codec) *cobra.Command {
 			accountFlags, err := strconv.ParseUint(flagsHexStr, 16, 64)
 			if err != nil {
 				return err
+			}
+			if !viper.GetBool(clientFlags.FlagOffline) {
+				acc, err := cliCtx.GetAccount(from)
+				if err != nil {
+					return err
+				}
+				appAccount, ok := acc.(types.NamedAccount)
+				if !ok {
+					return fmt.Errorf("unexpected account type")
+				}
+				flags := appAccount.GetFlags()
+
+				if flags == accountFlags {
+					return fmt.Errorf("the specified account flags is identical to its current value")
+				}
 			}
 			// build message
 			msg := account.NewSetAccountFlagsMsg(from, accountFlags)
@@ -85,10 +98,6 @@ func enableMemoCheckFlagCmd(cdc *wire.Codec) *cobra.Command {
 					return err
 				}
 			} else {
-				cliCtx := context.NewCLIContext().
-					WithCodec(cdc).
-					WithAccountDecoder(authcmd.GetAccountDecoder(cdc))
-
 				if err := cliCtx.EnsureAccountExistsFromAddr(from); err != nil {
 					return err
 				}
@@ -102,7 +111,10 @@ func enableMemoCheckFlagCmd(cdc *wire.Codec) *cobra.Command {
 				}
 				flags = appAccount.GetFlags()
 			}
-			flags = flags | scripts.TransferMemoCheckerFlag
+			flags, err = setFlagBits(flags, scripts.TransferMemoCheckerFlag)
+			if err != nil {
+				return err
+			}
 			// build message
 			msg := account.NewSetAccountFlagsMsg(from, flags)
 			err = msg.ValidateBasic()
@@ -144,10 +156,6 @@ func disableMemoCheckFlagCmd(cdc *wire.Codec) *cobra.Command {
 					return err
 				}
 			} else {
-				cliCtx := context.NewCLIContext().
-					WithCodec(cdc).
-					WithAccountDecoder(authcmd.GetAccountDecoder(cdc))
-
 				if err := cliCtx.EnsureAccountExistsFromAddr(from); err != nil {
 					return err
 				}
@@ -161,8 +169,10 @@ func disableMemoCheckFlagCmd(cdc *wire.Codec) *cobra.Command {
 				}
 				flags = appAccount.GetFlags()
 			}
-			invMemoCheck := ^scripts.TransferMemoCheckerFlag
-			flags = flags & invMemoCheck
+			flags, err = unsetFlagBits(flags, scripts.TransferMemoCheckerFlag)
+			if err != nil {
+				return err
+			}
 			// build message
 			msg := account.NewSetAccountFlagsMsg(from, flags)
 			err = msg.ValidateBasic()